+++ Sources/Renderer/Renderer.cpp
@@ -319,13 +319,13 @@ void Renderer::drawObject(Entity* object, const float* view, const float* projec
 		//	Load skeleton pose matrix list for vertex skinning calculation
 		std::vector<glm::mat4> pose = skeletonComp->getPose();
 		int loc = shaderToUse->getUniformLocation("skeletonPose");
-		if(loc >= 0) glUniformMatrix4fv(loc, pose.size(), FALSE, (float*) pose.data());
+		if(loc >= 0) glUniformMatrix4fv(loc, (int)pose.size(), FALSE, (float*) pose.data());
 
 		//	Load inverse bind pose matrix list for vertex skinning calculation
 		std::vector<glm::mat4> bind;
 		bind = skeletonComp->getInverseBindPose();
 		loc = shaderToUse->getUniformLocation("inverseBindPose");
-		if(loc >= 0) glUniformMatrix4fv(loc, bind.size(), FALSE, (float*) bind.data());
+		if(loc >= 0) glUniformMatrix4fv(loc, (int)bind.size(), FALSE, (float*) bind.data());
 	}
 
 	//	Draw mesh
